refactor: move all loaders
