+++ cognite/neat/_rules/exporters/_rules2dms.py
@@ -15,13 +15,7 @@ from cognite.client.data_classes.data_modeling import (
 )
 from cognite.client.exceptions import CogniteAPIError
 
-from cognite.neat._issues import IssueList
-from cognite.neat._issues.warnings import (
-    PrincipleOneModelOneSpaceWarning,
-    ResourceRetrievalWarning,
-)
-from cognite.neat._rules.models.dms import DMSRules, DMSSchema
-from cognite.neat._utils.cdf.loaders import (
+from cognite.neat._client.api.data_modeling_loaders import (
     ContainerLoader,
     DataModelingLoader,
     DataModelLoader,
@@ -29,6 +23,12 @@ from cognite.neat._utils.cdf.loaders import (
     SpaceLoader,
     ViewLoader,
 )
+from cognite.neat._issues import IssueList
+from cognite.neat._issues.warnings import (
+    PrincipleOneModelOneSpaceWarning,
+    ResourceRetrievalWarning,
+)
+from cognite.neat._rules.models.dms import DMSRules, DMSSchema
 from cognite.neat._utils.upload import UploadResult
 
 from ._base import CDFExporter
