--- cognite/neat/_utils/cdf/loaders/__init__.py
@@ -1,17 +1,0 @@
-from ._base import ResourceLoader
-from ._data_modeling import (
-    ContainerLoader,
-    DataModelingLoader,
-    DataModelLoader,
-    SpaceLoader,
-    ViewLoader,
-)
-
-__all__ = [
-    "DataModelingLoader",
-    "ContainerLoader",
-    "DataModelLoader",
-    "ResourceLoader",
-    "SpaceLoader",
-    "ViewLoader",
-]
